+++ comics/comics/walkoflife.py
@@ -9,16 +9,11 @@ class ComicData(ComicDataBase):
     language = 'no'
     url = 'http://walkoflife.nettserier.no/'
     start_date = '2008-06-23'
+    end_date = '2014-06-24'
+    active = False
     rights = 'Trond J. Stavås'
 
 
 class Crawler(CrawlerBase):
-    history_capable_date = '2008-06-23'
-    schedule = 'Tu,Fr'
-    time_zone = 'Europe/Oslo'
-
     def crawl(self, pub_date):
-        epoch = self.date_to_epoch(pub_date)
-        url = 'http://walkoflife.nettserier.no/_striper/walkoflife-%s.png' % (
-            epoch,)
-        return CrawlerImage(url)
+        pass  # Comic no longer published